+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added code to check for and warn when `pub get` should be run on a project
 - improved the ability for the pub get and pub update commands to select the
   correct pubspec.yaml file to operate on
+- updated some references from `sky_tools` to `flutter`
 
 ## 0.4.9
 - added UI to improve discoverability for quick fixes (`ctrl-1` / `alt-enter`)